Updated method for patching in recipes
There are two methods I know of for patching in and/or creating recipes. I initially created specific recipeDefs for all eltex apparel. I had commented out patches that added the recipes directly to the ThingDefs of the respective gear pieces. I swapped those methods and tweaked some cost values as well as market values. I believe the primary reason for using recipeDefs initially was that I could specify decimals for silver and gold amounts. Since these values are multiplied by 10 in-game, this allowed costs to be defined in intervals of 1 instead of 10. This no longer concerns me and I feel it was an over-analyzation.